Added integration test of the Employee Controller
This will start the server as it would behave in normal operation to test the Employee Endpoint
The data.sql file is to fill the database with data while running the integration test.
Lastly, the root log level was defined in both application.properties environments. That way you can change the log level independently of each other.