implemented the Service function to get one employee
Added the getOne function to the EmployeeService to get a single Employee by its ID.
Also added the EmployeeServiceImpl Tests to get a single Employee with an existing and non-existing result.
This also required to add a custom exception that is being thrown when we want to retrieve an Employee with an ID that doesn't exist.